+++ src/yafowil/widget/chosen/resources/widget.js
@@ -4,6 +4,10 @@ var yafowil_chosen = (function (exports, $) {
     class ChosenWidget {
         static initialize(context) {
             $('select.chosen', context).each(function (event) {
+                if (window.yafowil_array !== undefined &&
+                    window.yafowil_array.inside_template($(this))) {
+                    return;
+                }
                 new ChosenWidget($(this));
             });
         }
@@ -43,6 +47,15 @@ var yafowil_chosen = (function (exports, $) {
             }
         }
     }
+    function chosen_on_array_add(inst, context) {
+        ChosenWidget.initialize(context, true);
+    }
+    function register_array_subscribers() {
+        if (window.yafowil_array === undefined) {
+            return;
+        }
+        window.yafowil_array.on_array_event('on_add', chosen_on_array_add);
+    }
 
     $(function() {
         if (window.ts !== undefined) {
@@ -52,9 +65,11 @@ var yafowil_chosen = (function (exports, $) {
         } else {
             ChosenWidget.initialize();
         }
+        register_array_subscribers();
     });
 
     exports.ChosenWidget = ChosenWidget;
+    exports.register_array_subscribers = register_array_subscribers;
 
     Object.defineProperty(exports, '__esModule', { value: true });