+++ CHANGES.rst
@@ -4,6 +4,13 @@ Changes
 2.0 (unreleased)
 ----------------
 
+- Extend JS by ``chosen_on_array_add`` and ``register_array_subscribers``
+  functions to enable usage in ``yafowil.widget.array``.
+  [lenadax]
+
+- Prevent initialize of widget if part of array template.
+  [lenadax]
+
 - Rewrite JavaScript using ES6.
   [rnix]